+++ packages/@sanity/cli-build/src/_exports/_internal/manifest.ts
@@ -1,0 +1,23 @@
+export {
+  extractCoreAppManifest,
+  readIconFromPath,
+  resolveTitleUpdate,
+} from '../../actions/manifest/extractCoreAppManifest.js'
+export {extractManifest} from '../../actions/manifest/extractManifest.js'
+export {
+  extractManifestSchemaTypes,
+  extractWorkspaceManifest,
+} from '../../actions/manifest/extractWorkspaceManifest.js'
+export {
+  type CoreAppManifest,
+  type CreateManifest,
+  type CreateWorkspaceManifest,
+  CURRENT_WORKSPACE_SCHEMA_VERSION,
+  type DefaultWorkspaceSchemaId,
+  type ManifestSchemaType,
+  SANITY_WORKSPACE_SCHEMA_ID_PREFIX,
+  type StoredWorkspaceSchema,
+  type StudioManifest,
+  type WorkspaceSchemaId,
+} from '../../actions/manifest/types.js'
+export {MANIFEST_FILENAME, writeManifestFile} from '../../actions/manifest/writeManifestFile.js'