+++ packages/@sanity/cli-build/src/actions/schema/uploadSchemaToLexicon.ts
@@ -11,7 +11,8 @@ import {
 import {spinner} from '@sanity/cli-core/ux'
 import {type StudioManifest, type Workspace} from 'sanity'
 
-import {SCHEMA_API_VERSION} from '../../services/schemas.js'
+/** This version corresponds to the one used in the cli schema service but was copied here to avoid the dependency issue. */
+const SCHEMA_API_VERSION = 'v2025-03-01'
 
 const iconResolverPath = new URL('../manifest/iconResolver.js', import.meta.url).href